move chains.ts to shared
